+++ embassy-nrf/src/uarte.rs
@@ -24,15 +24,19 @@ use embassy_sync::waitqueue::AtomicWaker;
 // Re-export SVD variants to allow user to directly set values.
 pub use pac::uarte::vals::{Baudrate, ConfigParity as Parity};
 
-use crate::chip::{EASY_DMA_SIZE, FORCE_COPY_BUFFER_SIZE};
+use crate::chip::FORCE_COPY_BUFFER_SIZE;
 use crate::gpio::{self, AnyPin, DISCONNECTED, Pin as GpioPin, PselBits, SealedPin as _};
 use crate::interrupt::typelevel::Interrupt;
 use crate::pac::gpio::vals as gpiovals;
+use crate::pac::uarte::regs::RxMaxcnt;
 use crate::pac::uarte::vals;
 use crate::ppi::{AnyConfigurableChannel, ConfigurableChannel, Event, Ppi, Task};
 use crate::timer::{Frequency, Instance as TimerInstance, Timer};
 use crate::util::slice_in_ram_or;
 use crate::{interrupt, pac};
+
+/// The maximum buffer size (in bytes) that the UARTE EasyDMA can transfer in one operation.
+pub const DMA_SIZE: usize = crate::util::easy_dma_max!(RxMaxcnt, set_maxcnt, maxcnt);
 
 /// UARTE config.
 #[derive(Clone)]
@@ -471,7 +475,7 @@ impl<'d> UarteTx<'d> {
         }
 
         slice_in_ram_or(buffer, Error::BufferNotInRAM)?;
-        if buffer.len() > EASY_DMA_SIZE {
+        if buffer.len() > DMA_SIZE {
             return Err(Error::BufferTooLong);
         }
 
@@ -541,7 +545,7 @@ impl<'d> UarteTx<'d> {
         }
 
         slice_in_ram_or(buffer, Error::BufferNotInRAM)?;
-        if buffer.len() > EASY_DMA_SIZE {
+        if buffer.len() > DMA_SIZE {
             return Err(Error::BufferTooLong);
         }
 
@@ -701,7 +705,7 @@ impl<'d> UarteRx<'d> {
         if buffer.is_empty() {
             return Ok(());
         }
-        if buffer.len() > EASY_DMA_SIZE {
+        if buffer.len() > DMA_SIZE {
             return Err(Error::BufferTooLong);
         }
 
@@ -769,7 +773,7 @@ impl<'d> UarteRx<'d> {
         if buffer.is_empty() {
             return Ok(());
         }
-        if buffer.len() > EASY_DMA_SIZE {
+        if buffer.len() > DMA_SIZE {
             return Err(Error::BufferTooLong);
         }
 
@@ -886,7 +890,7 @@ impl<'d> UarteRx<'d> {
         if buffer.is_empty() {
             return Ok(0);
         }
-        if buffer.len() > EASY_DMA_SIZE {
+        if buffer.len() > DMA_SIZE {
             return Err(Error::BufferTooLong);
         }
 
@@ -945,7 +949,7 @@ impl<'d> UarteRx<'d> {
         if buffer.is_empty() {
             return Ok(0);
         }
-        if buffer.len() > EASY_DMA_SIZE {
+        if buffer.len() > DMA_SIZE {
             return Err(Error::BufferTooLong);
         }
 
@@ -1059,7 +1063,7 @@ impl<'d> UarteRxWithIdle<'d> {
         if buffer.is_empty() {
             return Ok(0);
         }
-        if buffer.len() > EASY_DMA_SIZE {
+        if buffer.len() > DMA_SIZE {
             return Err(Error::BufferTooLong);
         }
 
@@ -1133,7 +1137,7 @@ impl<'d> UarteRxWithIdle<'d> {
         if buffer.is_empty() {
             return Ok(0);
         }
-        if buffer.len() > EASY_DMA_SIZE {
+        if buffer.len() > DMA_SIZE {
             return Err(Error::BufferTooLong);
         }